Addprogram

+++ HolbertonDailyProjects/variables_if_else_while/how_To_AO.c
@@ -9,7 +9,7 @@ int main(void)
 	printf("Addition of op1 and op2: %d\n", op1 + op2);
 	printf("Subtraction of op1 and op2 %d\n", op1 - op2);
 	printf("Multiplication of op1 and op2 %d\n", op1 * op2);
-	printf("Division of op1 and op2 %d\n", op1 / op2);
+	printf("Division of op1 and op2 %f\n", (float)op1 / op2);
 
 	return 0;
 }
